pyproject.toml
[project]
name = "mcp-ssh-ops"
version = "0.1.0"
description = "MCP server for executing safe diagnostic commands over SSH"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.2.0",
    "asyncssh>=2.14.0",
    "pyyaml>=6.0",
]

[project.scripts]
mcp-ssh-ops = "mcp_ssh_ops.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
